refactor(specs): Simplify structure for specs
