+++ pkg/specs/aws-cis-1.2.yaml
