+++ pkg/specs/aws-cis-1.4.yaml
