+++ pkg/specs/docker-cis-1.6.0.yaml
