+++ pkg/specs/k8s-cis-1.23.yaml
