+++ pkg/specs/k8s-nsa-1.0.yaml
