+++ pkg/specs/k8s-pss-restricted-0.1.yaml
