+++ pkg/specs/loader.go
@@ -10,24 +10,40 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
-const ComplianceFolder = "compliance"
+// Loader access compliance specs
+type Loader interface {
+	GetSpecByName(name string) string
+}
+
+type specLoader struct {
+}
+
+// NewSpecLoader instansiate spec loader
+func NewSpecLoader() Loader {
+	return &specLoader{}
+}
+
+// GetSpecByName get spec name and return spec data
+func (sl specLoader) GetSpecByName(name string) string {
+	return GetSpec(name)
+}
 
 var (
-	//go:embed compliance
-	complainceFS embed.FS
+	//go:embed *.yaml
+	complianceFS embed.FS
 )
 
 var complianceSpecMap map[string]string
 
 // Load compliance specs
 func init() {
-	dir, _ := complainceFS.ReadDir(ComplianceFolder)
-	complianceSpecMap = make(map[string]string, 0)
+	dir, _ := complianceFS.ReadDir(".")
+	complianceSpecMap = make(map[string]string)
 	for _, r := range dir {
 		if !strings.Contains(r.Name(), ".yaml") {
 			continue
 		}
-		file, err := complainceFS.Open(fmt.Sprintf("%s/%s", ComplianceFolder, r.Name()))
+		file, err := complianceFS.Open(fmt.Sprintf("%s", r.Name()))
 		if err != nil {
 			panic(err)
 		}
